Added json handling for gun barrel data
Will be used to allow guns to be created with several barrels or exit points for attacks. Logic still needs to be added to do something with this.

Also this can be used purely for rendering purposes. In which the weapon will still only attack once. However, can have several visual attacks generated.