Implemented barrel logic, Reworked on shooter aim is handled
Shooter aim is very flaky for entities. Convering entity AIM to the rotation data used by the gun system breaks easily. A better solution needs to be looked at. As well something seems to be off with EularAngle.transform(pos) logic. Can't tell if its the bad math in the gun, but any attempt to transform a position failed.