Fixed cart textures not working
I really need to rework the JSON item icon system to not have 5 different ways to get icons. It needs to be a 1:1 system. My thoughts on how to do this is to define render types in the JSON data. That then are mapped to the keys at load time. Instead of generating keys, checking if they exist, and then getting an icon.